Move Timewise.Part and Timewise.Measure to own files
